# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthenticationResponse": {
        "description": "登录或注册成功后返回的 Bearer Token 与账号信息。",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/CurrentUserResponse"
          }
        },
        "required": ["access_token", "user"],
        "title": "AuthenticationResponse",
        "type": "object"
      },
...
      "CandidateCitationResponse": {
        "description": "由后端从已核验格式中立题录渲染出的单个正式引用。",
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/CitationFormat"
          },
          "text": {
            "description": "可直接复制或导出的正式引用文本",
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": ["candidate_id", "format", "text"],
        "title": "CandidateCitationResponse",
        "type": "object"
      },
...
      "CitationFormat": {
        "description": "当前可从同一份已核验题录导出的引用格式。",
        "enum": ["gb_t_7714_2015_numeric", "apa_7", "mla_9", "chicago_author_date", "bibtex"],
        "title": "CitationFormat",
        "type": "string"
      },
...
      "CreateWorkspaceRequest": {
        "description": "新建研究工作区的最小输入，不允许客户端指定所有者或状态。",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": ["name"],
        "title": "CreateWorkspaceRequest",
        "type": "object"
      },
      "CurrentUserResponse": {
        "description": "可安全返回给前端的当前账号信息。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": ["id", "email", "display_name", "created_at"],
        "title": "CurrentUserResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "description": "本地账号密码登录输入。",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": ["email", "password"],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "RegisterRequest": {
        "description": "创建本地账号所需的最小信息。",
        "properties": {
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          }
        },
        "required": ["email", "password", "display_name"],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowStageDisplay": {
        "description": "工作区研究阶段的中文展示元数据，不作为客户端提交输入。",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": ["label", "description"],
        "title": "WorkflowStageDisplay",
        "type": "object"
      },
      "WorkspaceListResponse": {
        "description": "工作区切换器使用的游标分页结果。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": ["items"],
        "title": "WorkspaceListResponse",
        "type": "object"
      },
      "WorkspaceResponse": {
        "description": "工作区列表和详情共用的可展示字段。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "research_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Question"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_stage": {
            "$ref": "#/components/schemas/WorkspaceWorkflowStage"
          },
          "workflow_stage_display": {
            "$ref": "#/components/schemas/WorkflowStageDisplay"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "research_question",
          "status",
          "workflow_stage",
          "workflow_stage_display",
          "created_at",
          "updated_at"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      },
      "WorkspaceWorkflowStage": {
        "description": "工作区在研究主流程中的阶段；英文值会写入数据库和 API。",
        "enum": [
          "draft",
          "analyzing",
          "plan_review",
          "retrieving",
          "screening",
          "collection_building",
          "researching",
          "failed"
        ],
        "title": "WorkspaceWorkflowStage",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "面向学术文献检索与 RAG 研究工作区的后端服务。",
    "title": "Academic Search API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "验证本地账号密码并签发新的短生命周期访问令牌。",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "使用邮箱密码登录",
        "tags": ["认证"]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "让前端在刷新后确认 Bearer Token 对应的登录用户。",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "获取当前登录账号",
        "tags": ["认证"]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "创建账号并立即返回访问令牌，邮箱验证将在后续独立接入。",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "注册本地账号",
        "tags": ["认证"]
      }
    },
    "/api/v1/collections": {
      "get": {
        "description": "供工作区切换器按需加载，默认只显示活动工作区。",
        "operationId": "list_workspaces_api_v1_collections_get",
        "parameters": [
          {
            "description": "是否包含已归档工作区",
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "description": "是否包含已归档工作区",
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "description": "按工作区名称或当前研究阶段搜索",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按工作区名称或当前研究阶段搜索",
              "title": "Q"
            }
          },
          {
            "description": "上一页返回的不透明分页游标",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "上一页返回的不透明分页游标",
              "title": "Cursor"
            }
          },
          {
            "description": "本次最多返回的工作区数量",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "本次最多返回的工作区数量",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "搜索我的研究工作区",
        "tags": ["研究工作区"]
      },
      "post": {
        "description": "创建一个空工作区；研究问题将在进入工作区后另行确认。",
        "operationId": "create_workspace_api_v1_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "创建研究工作区",
        "tags": ["研究工作区"]
      }
    },
...
    "/api/v1/collections/{collection_id}/search-runs/{search_run_id}/candidates/{candidate_id}/citation": {
      "get": {
        "description": "从 Redis 候选快照的 `ready` 格式中立题录渲染一种可复制引用。",
        "operationId": "render_candidate_citation_api_v1_collections__collection_id__search_runs__search_run_id__candidates__candidate_id__citation_get",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "search_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Search Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "description": "需要渲染的正式引用格式",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CitationFormat",
              "default": "gb_t_7714_2015_numeric",
              "description": "需要渲染的正式引用格式"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateCitationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "生成候选的正式引用",
        "tags": ["正式引用"]
      }
    },
...
  }
}